feat(api): chat 代码拆分到 utils 中
